# model/src/main/java/com/showtime/model/view/generator/GeneratorView.java
package com.showtime.model.view.generator;

import com.showtime.model.entity.generator.Generator;
import io.swagger.annotations.ApiModel;

import java.io.Serializable;

/**
 * <b><code>GeneratorView</code></b>
 * <p/>
 * Generator的具体实现
 * <p/>
 * <b>Creation Time:</b> Fri Jul 28 16:32:21 CST 2017.
 *
 * @author qinJianLun
 * @version 1.0.0
 * @since model 1.0.0
 */
@ApiModel
public class GeneratorView extends Generator implements Serializable {

    /**
     * The constant serialVersionUID.
     */
    private static final long serialVersionUID = -1L;

}