# rs/generated/swagger-ui/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "RESTful API文档",
    "version" : "v1.0",
    "title" : "RESTful API文档",
    "termsOfService" : "http://www.showtime.com/",
    "contact" : {
      "name" : "showtime",
      "url" : "http://www.showtime.com",
      "email" : "594470882@qq.com"
    }
  },
  "basePath" : "/rs",
...
  "schemes" : [ "http", "https" ],
  "paths" : {
...
    "/api/v1/generators" : {
      "get" : {
        "tags" : [ "后缀生成接口" ],
        "summary" : "获取后缀生成列表",
        "description" : "通过查询条件获取后缀生成列表",
        "operationId" : "getGenerators",
        "produces" : [ "application/json;charset=UTF-8" ],
        "parameters" : [ {
          "name" : "enabled",
          "in" : "query",
          "description" : "状态",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "pageNumber",
          "in" : "query",
          "description" : "页数",
          "required" : false,
          "type" : "integer",
          "default" : 0,
          "format" : "int32"
        }, {
          "name" : "pageSize",
          "in" : "query",
          "description" : "每页加载量",
          "required" : false,
          "type" : "integer",
          "default" : 10,
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful request"
          },
          "500" : {
            "description" : "internal server error"
          }
        }
      },
      "post" : {
        "tags" : [ "后缀生成接口" ],
        "summary" : "创建后缀生成",
        "description" : "创建一个后缀生成",
        "operationId" : "createGenerator",
        "produces" : [ "application/json;charset=UTF-8" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "后缀生成",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/GeneratorView"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "successful operation",
            "headers" : {
              "location" : {
                "type" : "string",
                "description" : "URL of new created resource"
              }
            }
          },
          "409" : {
            "description" : "conflict"
          },
          "500" : {
            "description" : "internal server error"
          }
        }
      },
      "delete" : {
        "tags" : [ "后缀生成接口" ],
        "summary" : "删除后缀生成",
        "description" : "批量删除后缀生成",
        "operationId" : "deleteGenerators",
        "produces" : [ "application/json;charset=UTF-8" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "后缀生成ids，样例 - 1,2,3",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful request"
          },
          "500" : {
            "description" : "internal server error"
          }
        }
      }
    },
    "/api/v1/generators/{id}" : {
      "get" : {
        "tags" : [ "后缀生成接口" ],
        "summary" : "获取单个后缀生成",
        "description" : "通过id获取后缀生成",
        "operationId" : "getGenerator",
        "produces" : [ "application/json;charset=UTF-8" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "后缀生成id",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful request"
          },
          "500" : {
            "description" : "internal server error"
          }
        }
      },
      "put" : {
        "tags" : [ "后缀生成接口" ],
        "summary" : "更新后缀生成",
        "description" : "更新后缀生成信息",
        "operationId" : "updateGenerators",
        "produces" : [ "application/json;charset=UTF-8" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "后缀生成id",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "后缀生成信息",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/GeneratorView"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful"
          },
          "404" : {
            "description" : "not found"
          },
          "409" : {
            "description" : "conflict"
          },
          "500" : {
            "description" : "internal Server Error"
          }
        }
      },
      "delete" : {
        "tags" : [ "后缀生成接口" ],
        "summary" : "删除后缀生成",
        "description" : "通过id删除后缀生成",
        "operationId" : "deleteGenerator",
        "produces" : [ "application/json;charset=UTF-8" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "后缀生成id",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful request"
          },
          "500" : {
            "description" : "internal server error"
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "GeneratorView" : {
      "type" : "object",
      "required" : [ "createTime", "updateTime" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "createTime" : {
          "type" : "integer",
          "format" : "int64"
        },
        "updateTime" : {
          "type" : "integer",
          "format" : "int64"
        },
        "enabled" : {
          "type" : "integer",
          "format" : "int32"
        },
        "auditors" : {
          "type" : "string"
        },
        "auditStatus" : {
          "type" : "integer",
          "format" : "int32"
        },
        "auditContent" : {
          "type" : "string"
        },
        "auditTime" : {
          "type" : "integer",
          "format" : "int64"
        },
        "name" : {
          "type" : "string"
        },
        "numberSuffix" : {
          "type" : "string"
        }
      }
    },
...
  }
}